Extract similar code into parent classes
