--- decidim-assemblies/app/models/decidim/assembly_role_config/admin.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module AssemblyRoleConfig
-    class Admin < Base
-      def component_is_whitelisted?(_manifest)
-        true
-      end
-    end
-  end
-end
