--- decidim-assemblies/app/models/decidim/assembly_role_config/moderator.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module AssemblyRoleConfig
-    class Moderator < Base
-      def component_is_whitelisted?(_manifest)
-        false
-      end
-    end
-  end
-end
