--- decidim-assemblies/app/models/decidim/assembly_role_config/valuator.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module AssemblyRoleConfig
-    class Valuator < Base
-      def component_is_whitelisted?(manifest)
-        manifest.to_sym == :proposals
-      end
-    end
-  end
-end
