+++ decidim-conferences/app/models/decidim/conference.rb
@@ -124,23 +124,6 @@ module Decidim
       roles.where(role: role_name)
     end
 
-    def user_role_config_for(user, role_name)
-      case role_name.to_sym
-      when :organization_admin
-        Decidim::ConferenceRoleConfig::Admin.new(user)
-      when :admin # conference admin
-        Decidim::ConferenceRoleConfig::ConferenceAdmin.new(user)
-      when :valuator
-        Decidim::ConferenceRoleConfig::Valuator.new(user)
-      when :moderator
-        Decidim::ConferenceRoleConfig::Moderator.new(user)
-      when :collaborator
-        Decidim::ConferenceRoleConfig::Collaborator.new(user)
-      else
-        Decidim::ConferenceRoleConfig::NullObject.new(user)
-      end
-    end
-
     # Allow ransacker to search for a key in a hstore column (`title`.`en`)
     ransacker :title do |parent|
       Arel::Nodes::InfixOperation.new("->>", parent.table[:title], Arel::Nodes.build_quoted(I18n.locale.to_s))
