--- decidim-conferences/app/models/decidim/conference_role_config/base.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ConferenceRoleConfig
-    class Base < Decidim::UserRoleConfig::Base
-    end
-  end
-end
