--- decidim-conferences/app/models/decidim/conference_role_config/null_object.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ConferenceRoleConfig
-    class NullObject < Base
-      def component_is_whitelisted?(_manifest)
-        false
-      end
-    end
-  end
-end
