+++ decidim-core/app/models/decidim/user_role_config/admin.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 module Decidim
-  module AssemblyRoleConfig
-    class NullObject < Base
+  module ParticipatorySpaceRoleConfig
+    class Admin < Base
       def component_is_whitelisted?(_manifest)
-        false
+        true
       end
     end
   end
