+++ decidim-core/app/models/decidim/user_role_config/collaborator.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Decidim
-  module ConferenceRoleConfig
-    class Admin < Base
+  module ParticipatorySpaceRoleConfig
+    class Collaborator < Base
       def component_is_whitelisted?(_manifest)
         true
       end
