+++ decidim-core/app/models/decidim/user_role_config/participatory_space_admin.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Decidim
-  module AssemblyRoleConfig
-    class AssemblyAdmin < Base
+  module ParticipatorySpaceRoleConfig
+    class ParticipatorySpaceAdmin < Base
       def component_is_whitelisted?(_manifest)
         true
       end
