+++ decidim-core/app/models/decidim/user_role_config/valuator.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Decidim
-  module ConferenceRoleConfig
+  module ParticipatorySpaceRoleConfig
     class Valuator < Base
       def component_is_whitelisted?(manifest)
         manifest.to_sym == :proposals
