--- decidim-participatory_processes/app/models/decidim/participatory_process_role_config/admin.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ParticipatoryProcessRoleConfig
-    class Admin < Base
-      def component_is_whitelisted?(_manifest)
-        true
-      end
-    end
-  end
-end
