--- decidim-participatory_processes/app/models/decidim/participatory_process_role_config/base.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ParticipatoryProcessRoleConfig
-    class Base < Decidim::UserRoleConfig::Base
-    end
-  end
-end
