--- decidim-participatory_processes/app/models/decidim/participatory_process_role_config/collaborator.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ParticipatoryProcessRoleConfig
-    class Collaborator < Base
-      def component_is_whitelisted?(_manifest)
-        true
-      end
-    end
-  end
-end
