--- decidim-participatory_processes/app/models/decidim/participatory_process_role_config/null_object.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ParticipatoryProcessRoleConfig
-    class NullObject < Base
-      def component_is_whitelisted?(_manifest)
-        false
-      end
-    end
-  end
-end
