--- decidim-participatory_processes/app/models/decidim/participatory_process_role_config/valuator.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module ParticipatoryProcessRoleConfig
-    class Valuator < Base
-      def component_is_whitelisted?(manifest)
-        manifest.to_sym == :proposals
-      end
-    end
-  end
-end
